Probably this can be a size_t.  Unsigned is fine, and greater than max size_t triangulations will surely never finish.

+++ utils/tricensus.cpp
@@ -89,7 +89,7 @@ std::unique_ptr<std::ostream> dumpStream;
 unsigned long totPairings = 0;
 
 // Variables used for output.
-long long nSolns;
+size_t nSolns;
 bool textOutput;
 std::ofstream sigStream;
